requirements.txt
flask
pandas
scikit-learn
joblib
pyyaml
dvc


# optional utils
matplotlib
seaborn